dump/requirements.txt
flask[async]>=2.0.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
blinker==1.9.0
click==8.1.8
fastapi==0.115.12
Flask==3.1.0
frozenlist==1.5.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.2.0
numpy==2.2.4
pandas==2.2.3
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
typing_extensions==4.12.2
tzdata==2025.2
uvicorn==0.34.0
Werkzeug==3.1.3
xmltodict==0.14.2
yarl==1.18.3
